added a few things i think

+++ src/index.ts
@@ -4,39 +4,84 @@
  */
 import dotenv from 'dotenv';
 import commands from './commands/index';
-import { app, CLIENT, io, normalized, server } from './util';
+import { app, CLIENT, inBotChannel, io, normalized, server } from './util';
 import slashBuilder from './commands/builder'
 import { tactics } from './types';
 import { AutocompleteInteraction } from 'discord.js';
 import path from 'path';
 import { addTactic } from './commands/game state/add_tactic';
+import { removeTactic } from './commands/game state/remove_tactic';
+import { nextTurn } from './commands/game state/next_turn';
+import { roll } from './commands/util/roll';
 dotenv.config();
 
-CLIENT.on('ready', () => { console.log("Ready!"); slashBuilder(); });
-CLIENT.on('messageCreate', message => {
-  if (message.author.id == '269333441982496769' && message.content.startsWith("!d")) {// aint no fuckin way i'm giving eval to anyone else
+const TACTIC_ADD_REGEX = /([^ \/@|]+)(?:\/([^ \/@|]+))?(?:\/([^ \/@|]+))?(?:@([^ \/@|]+))?(?:\|([^ \/@|]+))?/ // wow that looks ugly
+const TACTIC_REMOVE_REGEX = /([^ \/@|]+)(?:\/([^ \/@|]+))?(?:\/([^ \/@|]+))?/
+
+CLIENT.on('clientReady', () => { console.log("Ready!"); slashBuilder(); });
+// CLIENT.on('ready', async () => {
+//   const msg = (await CLIENT.guilds.cache.get("1210373907174006814")?.channels.fetch("1212180105460326520"))
+//   if (msg?.isSendable()) msg.send("let's do this bitchesssssss")
+// });
+CLIENT.on('messageCreate', async message => {
+  if (message.content.startsWith("?r ") || message.content.startsWith("?roll ")) {
+    const [reply, success] = roll(message.content.slice(2).trim(), message.author.id);
+    const item = await message.channel.send({
+      content: typeof reply === 'string' ? reply : `<@${message.author.id}>`,
+      embeds: typeof reply === 'string' ? [] : [reply],
+      allowedMentions: { users: [] }
+    });
+    if (!success) setTimeout(() => item.delete(), 10_000);
+  }
+  if (message.author.id == '269333441982496769' && message.content.startsWith("?d ")) {// aint no fuckin way i'm giving eval to anyone else
     const evals = eval(`(()=>{${message.content.slice(2).trim()}})()`);
     if (typeof evals !== 'undefined') {
       message.channel.send(evals.toString())
         .then(m => setTimeout(() => m.delete(), 10000));
     }
   }
-  if (message.author.id == '269333441982496769' && message.content.startsWith("!a")) { // we'll see if it catches on with others
+  if (message.author.id == '269333441982496769' && message.content.startsWith("?a ")) { // we'll see if it catches on with others
+    // TODO: from here on out it's shorthand syntax which is why this check exists, refactor it somehow
+    if (!global.game.currentPlayer) return message.react("❌");
+    const args = message.content.split(" ").splice(1);
+    // always assumed that the user is the person currently going
+    if (!args.length) {
+      message.channel.send("Usage: ?a (name/target/turns/@input|user)").then(m => setTimeout(() => m.delete(), 10000));
+      return;
+    }
+    const syntaxes: [string, boolean, string][] = args.map(syntax => { //  TODO: validate split
+      if (!TACTIC_ADD_REGEX.test(syntax)) return [syntax, false, "Invalid syntax."];
+      const [_, name, target, turns, input, user] = syntax.match(TACTIC_ADD_REGEX)!.map(x => x ?? null);
+      const [message, success] = addTactic(name, user, target, Number(turns || 1), input);
+      return [syntax, success, message];
+    })
+    if (syntaxes.length > 1 && !syntaxes.every(([_, success]) => success)) {
+      const failedSyntaxes = syntaxes.filter(([_, success]) => !success).map(([syntax, _, message]) => `${syntax} (${message})`).join(", ");
+      message.channel.send(`The following syntaxes failed: ${failedSyntaxes}`)
+        .then(m => setTimeout(() => m.delete(), 10000));
+    }
+    if (syntaxes.some(([_, success]) => success)) message.react("✅");
+  }
+  if (message.author.id == '269333441982496769' && message.content.startsWith("!n")) {
+    const messageLogs = nextTurn();
+    message.channel.send(`\`\`\`\n${messageLogs.join("\n")}\n\`\`\``);
+  }
+  if (message.author.id == '269333441982496769' && message.content.startsWith("!rm ")) { // copied the !a
     // TODO: from here on out it's shorthand syntax which is why this check exists, refactor it somehow
     if (!global.game.currentPlayer) return message.react("❌");
     const args = message.content.split(" ").splice(1);
     // always assumed that the user is the person currently going
     if (!args) {
-      message.channel.send("Usage: !a (name/target/turns/@input)").then(m => setTimeout(() => m.delete(), 10000));
+      message.channel.send("Usage: !rm (name/target/user)").then(m => setTimeout(() => m.delete(), 10000));
       return;
     }
     const syntaxes = args.map(syntax => {
-      const [rest, input] = syntax.split("@");
-      const [name, target, turns] = rest.split("/");
-      const [message, success] = addTactic(name, null, target || null, Number(turns || 1), input || null);
+      if (!TACTIC_REMOVE_REGEX.test(syntax)) return [syntax, false, "Invalid syntax."];
+      const [_, name, target, user] = syntax.match(TACTIC_REMOVE_REGEX)!.map(x => x ?? null);
+      const [message, success] = removeTactic(name, target, user);
       return [syntax, success, message];
     })
-    if (syntaxes.length > 1 || !syntaxes.every(([_, success]) => success)) {
+    if (syntaxes.length > 1 && !syntaxes.every(([_, success]) => success)) {
       const failedSyntaxes = syntaxes.filter(([_, success]) => !success).map(([syntax, _, message]) => `${syntax} (${message})`).join(", ");
       message.channel.send(`The following syntaxes failed: ${failedSyntaxes}`)
         .then(m => setTimeout(() => m.delete(), 10000));
@@ -74,6 +119,8 @@ async function onAutocomplete(interaction: AutocompleteInteraction) {
 process.on('uncaughtException', console.error);
 CLIENT.login(process.env.TOKEN);
 
-app.get('/', (_, res) => res.sendFile(path.join(__dirname, 'index.html')));
-io.on('connection', socket => socket.emit('contentUpdate', global.game, tactics));
-server.listen(3000, () => console.log('Listening on http://localhost:3000'));
+if (process.argv[2] !== "--no-screen") {
+  app.get('/', (_, res) => res.sendFile(path.join(__dirname, 'index.html')));
+  io.on('connection', socket => socket.emit('contentUpdate', global.game, tactics));
+  server.listen(3000, () => console.log('Listening on http://localhost:3000'));
+}
